+++ server/ProjectPokemon/ProjectPokemon/Models/Database/Repositories/MovementRepository.cs
@@ -1,7 +1,7 @@
 ﻿using ProjectPokemon.Models.Database.Entities;
 
 namespace ProjectPokemon.Models.Database.Repositories; 
-public class MovementRepository : BaseRepository<Movement, long> {
+public class MovementRepository : BaseRepository<Movement, int> {
     public MovementRepository(PokemonDbContext context) : base(context) {
     }
 }
