+++ server/ProjectPokemon/ProjectPokemon/Models/Database/Repositories/PokemonTeamRepository.cs
@@ -1,6 +1,6 @@
 ﻿using ProjectPokemon.Models.Database.Entities;
 
 namespace ProjectPokemon.Models.Database.Repositories; 
-public class PokemonTeamRepository : BaseRepository<PokemonTeam, long> {
+public class PokemonTeamRepository : BaseRepository<PokemonTeam, int> {
     public PokemonTeamRepository(PokemonDbContext context) : base(context) { }
 }
